Build docker image
Since the project depends on several libclang patches which are not yet
merged in the libclang
(https://www.mail-archive.com/cfe-commits@cs.uiuc.edu/msg95414.html,
http://lists.llvm.org/pipermail/cfe-commits/Week-of-Mon-20140428/104048.html),
the easiest is to use the provided pre-build binary and to build and run
the software inside a container.

Unfortunatelly, there doesn't seem to be any workaround in using
upstream libraries at this point.